Move CryptographyProvider and RCC into private headers as well.
