+++ sdk/keyvault/azure-security-keyvault-keys/inc/azure/keyvault/keys/cryptography/cryptography_client.hpp
@@ -11,24 +11,35 @@
 
 #include "../src/private/keyvault_protocol.hpp"
 
+#include <azure/core/context.hpp>
 #include <azure/core/io/body_stream.hpp>
 
 #include "azure/keyvault/keys/cryptography/cryptography_client_options.hpp"
-#include "azure/keyvault/keys/cryptography/cryptography_provider.hpp"
+#include "azure/keyvault/keys/cryptography/decrypt_parameters.hpp"
+#include "azure/keyvault/keys/cryptography/decrypt_result.hpp"
 #include "azure/keyvault/keys/cryptography/encrypt_parameters.hpp"
 #include "azure/keyvault/keys/cryptography/encrypt_result.hpp"
-#include "azure/keyvault/keys/cryptography/remote_cryptography_client.hpp"
+#include "azure/keyvault/keys/cryptography/key_wrap_algorithm.hpp"
 #include "azure/keyvault/keys/cryptography/sign_result.hpp"
+#include "azure/keyvault/keys/cryptography/signature_algorithm.hpp"
+#include "azure/keyvault/keys/cryptography/unwrap_result.hpp"
+#include "azure/keyvault/keys/cryptography/verify_result.hpp"
 #include "azure/keyvault/keys/cryptography/wrap_result.hpp"
 
 #include <memory>
 #include <string>
+#include <vector>
 
 namespace Azure {
   namespace Security {
     namespace KeyVault {
       namespace Keys {
         namespace Cryptography {
+
+  namespace _detail {
+    struct CryptographyProvider;
+    struct RemoteCryptographyClient;
+  } // namespace _detail
 
   /**
    * @brief A client used to perform cryptographic operations with Azure Key Vault keys.
@@ -78,6 +89,12 @@ namespace Azure {
     }
 
     /**
+     * @brief Destructs `%CryptographyClient`.
+     *
+     */
+    ~CryptographyClient();
+
+    /**
      * @brief Encrypts plaintext.
      *
      * @param parameters An #EncryptParameters containing the data to encrypt and other parameters
@@ -110,8 +127,8 @@ namespace Azure {
      * @param key The key to encrypt.
      * @param context A #Azure::Core::Context to cancel the operation.
      * @return The result of the wrap operation. The returned #WrapResult contains the wrapped key
-     * along with all other information needed to unwrap it. This information should be stored with
-     * the wrapped key.
+     * along with all other information needed to unwrap it. This information should be stored
+     * with the wrapped key.
      */
     WrapResult WrapKey(
         KeyWrapAlgorithm algorithm,
@@ -124,8 +141,8 @@ namespace Azure {
      * @param algorithm The #KeyWrapAlgorithm to use.
      * @param encryptedKey The encrypted key.
      * @param context A #Azure::Core::Context to cancel the operation.
-     * @return The result of the unwrap operation. The returned #UnwrapResult contains the key along
-     * with information regarding the algorithm and key used to unwrap it.
+     * @return The result of the unwrap operation. The returned #UnwrapResult contains the key
+     * along with information regarding the algorithm and key used to unwrap it.
      */
     UnwrapResult UnwrapKey(
         KeyWrapAlgorithm algorithm,
@@ -140,8 +157,8 @@ namespace Azure {
      * must be compatable with the specified algorithm.
      * @param context A #Azure::Core::Context to cancel the operation.
      * @return The result of the sign operation. The returned #SignResult contains the signature
-     * along with all other information needed to verify it. This information should be stored with
-     * the signature.
+     * along with all other information needed to verify it. This information should be stored
+     * with the signature.
      */
     SignResult Sign(
         SignatureAlgorithm algorithm,
@@ -155,8 +172,8 @@ namespace Azure {
      * @param data The data to sign.
      * @param context A #Azure::Core::Context to cancel the operation.
      * @return The result of the sign operation. The returned #SignResult contains the signature
-     * along with all other information needed to verify it. This information should be stored with
-     * the signature.
+     * along with all other information needed to verify it. This information should be stored
+     * with the signature.
      */
     SignResult SignData(
         SignatureAlgorithm algorithm,
@@ -170,8 +187,8 @@ namespace Azure {
      * @param data The data to sign.
      * @param context A #Azure::Core::Context to cancel the operation.
      * @return The result of the sign operation. The returned #SignResult contains the signature
-     * along with all other information needed to verify it. This information should be stored with
-     * the signature.
+     * along with all other information needed to verify it. This information should be stored
+     * with the signature.
      */
     SignResult SignData(
         SignatureAlgorithm algorithm,
@@ -181,8 +198,8 @@ namespace Azure {
     /**
      * @brief Verifies the specified signature.
      *
-     * @param algorithm The #SignatureAlgorithm to use. This must be the same algorithm used to sign
-     * the digest.
+     * @param algorithm The #SignatureAlgorithm to use. This must be the same algorithm used to
+     * sign the digest.
      * @param digest The pre-hashed digest corresponding to the signature. The hash algorithm used
      * to compute the digest must be compatable with the specified algorithm.
      * @param signature The signature to verify.
@@ -199,8 +216,8 @@ namespace Azure {
     /**
      * @brief Verifies the specified signature.
      *
-     * @param algorithm The #SignatureAlgorithm to use. This must be the same algorithm used to sign
-     * the data.
+     * @param algorithm The #SignatureAlgorithm to use. This must be the same algorithm used to
+     * sign the data.
      * @param data The data corresponding to the signature.
      * @param signature The signature to verify.
      * @param context A #Azure::Core::Context to cancel the operation.
@@ -216,8 +233,8 @@ namespace Azure {
     /**
      * @brief Verifies the specified signature.
      *
-     * @param algorithm The #SignatureAlgorithm to use. This must be the same algorithm used to sign
-     * the data.
+     * @param algorithm The #SignatureAlgorithm to use. This must be the same algorithm used to
+     * sign the data.
      * @param data The data corresponding to the signature.
      * @param signature The signature to verify.
      * @param context A #Azure::Core::Context to cancel the operation.
