+++ sdk/keyvault/azure-security-keyvault-keys/src/cryptography/cryptography_client.cpp
@@ -9,11 +9,11 @@
 #include <azure/core/http/policies/policy.hpp>
 
 #include "azure/keyvault/keys/cryptography/cryptography_client.hpp"
-#include "azure/keyvault/keys/cryptography/cryptography_provider.hpp"
-#include "azure/keyvault/keys/cryptography/remote_cryptography_client.hpp"
 #include "azure/keyvault/keys/key_operation.hpp"
 
+#include "../private/cryptography_provider.hpp"
 #include "../private/local_cryptography_provider_factory.hpp"
+#include "../private/remote_cryptography_client.hpp"
 
 #include <memory>
 #include <string>
@@ -53,6 +53,8 @@ inline std::vector<uint8_t> CreateDigest(
   return hashAlgorithm->Final(data.data(), data.size());
 }
 } // namespace
+
+CryptographyClient::~CryptographyClient() = default;
 
 void CryptographyClient::Initialize(std::string const&, Azure::Core::Context const& context)
 {
