+++ sdk/keyvault/azure-security-keyvault-keys/src/private/cryptography_provider.hpp
